Fix the "fs" issue by moving "getCustomer" function into a different file
