+++ src/layouts/demo/demoLayoutSSG.ts
@@ -10,6 +10,7 @@ import {
   getCustomer,
   getStaticAirtableDataset,
 } from '@/modules/core/airtable/getAirtableDataset';
+import { getCustomer } from '@/modules/core/airtable/dataset';
 import prepareAndSanitizeAirtableDataset from '@/modules/core/airtable/prepareAndSanitizeAirtableDataset';
 import { AirtableSchema } from '@/modules/core/airtable/types/AirtableSchema';
 import { RawAirtableRecordsSet } from '@/modules/core/airtable/types/RawAirtableRecordsSet';
